refactor: move query eval to runtime/data
