+++ interpreter/interpreter.go
@@ -2024,178 +2024,15 @@ func (i *Interpreter) evalQuery(q *parser.QueryExpr) (any, error) {
 	if err != nil {
 		return nil, err
 	}
-	var list []any
-	switch v := src.(type) {
-	case []any:
-		list = v
-	case *data.Group:
-		list = v.Items
-	default:
-		return nil, fmt.Errorf("query source must be list, got %T", src)
-	}
 
 	child := types.NewEnv(i.env)
 	old := i.env
 	i.env = child
 	defer func() { i.env = old }()
 
-	opts := data.QueryOptions{}
-
-	setEnv := func(item any) {
-		if m, ok := item.(map[string]any); ok && m["__join__"] == true {
-			for k, v := range m {
-				if k == "__join__" {
-					continue
-				}
-				child.SetValue(k, v, true)
-			}
-		} else {
-			child.SetValue(q.Var, item, true)
-		}
-	}
-
-	for _, f := range q.Froms {
-		srcVal, err := i.evalExpr(f.Src)
-		if err != nil {
-			return nil, err
-		}
-		var joinList []any
-		switch vv := srcVal.(type) {
-		case []any:
-			joinList = vv
-		case *data.Group:
-			joinList = vv.Items
-		default:
-			return nil, fmt.Errorf("join source must be list, got %T", srcVal)
-		}
-
-		fc := f
-		opts.Joins = append(opts.Joins, data.Join{
-			Items: joinList,
-			Merge: func(left, right any) (any, error) {
-				m := map[string]any{"__join__": true}
-				if lm, ok := left.(map[string]any); ok && lm["__join__"] == true {
-					for k, v := range lm {
-						if k == "__join__" {
-							continue
-						}
-						m[k] = v
-					}
-				} else {
-					m[q.Var] = left
-				}
-				m[fc.Var] = right
-				return m, nil
-			},
-		})
-	}
-
-	for _, j := range q.Joins {
-		srcVal, err := i.evalExpr(j.Src)
-		if err != nil {
-			return nil, err
-		}
-		var joinList []any
-		switch vv := srcVal.(type) {
-		case []any:
-			joinList = vv
-		case *data.Group:
-			joinList = vv.Items
-		default:
-			return nil, fmt.Errorf("join source must be list, got %T", srcVal)
-		}
-
-		jc := j // capture
-		opts.Joins = append(opts.Joins, data.Join{
-			Items: joinList,
-			On: func(left, right any) (bool, error) {
-				setEnv(left)
-				child.SetValue(jc.Var, right, true)
-				cond, err := i.evalExpr(jc.On)
-				if err != nil {
-					return false, err
-				}
-				return truthy(cond), nil
-			},
-			Merge: func(left, right any) (any, error) {
-				m := map[string]any{"__join__": true}
-				if lm, ok := left.(map[string]any); ok && lm["__join__"] == true {
-					for k, v := range lm {
-						if k == "__join__" {
-							continue
-						}
-						m[k] = v
-					}
-				} else {
-					m[q.Var] = left
-				}
-				m[jc.Var] = right
-				return m, nil
-			},
-			Left:  jc.Side != nil && (*jc.Side == "left" || *jc.Side == "outer"),
-			Right: jc.Side != nil && (*jc.Side == "right" || *jc.Side == "outer"),
-		})
-	}
-
-	if q.Where != nil {
-		opts.Where = func(item any) (bool, error) {
-			setEnv(item)
-			cond, err := i.evalExpr(q.Where)
-			if err != nil {
-				return false, err
-			}
-			return truthy(cond), nil
-		}
-	}
-
-	if q.Group != nil {
-		opts.GroupBy = func(item any) (any, error) {
-			setEnv(item)
-			return i.evalExpr(q.Group.Expr)
-		}
-		opts.SelectGroup = func(g *data.Group) (any, error) {
-			child.SetValue(q.Group.Name, g, true)
-			return i.evalExpr(q.Select)
-		}
-	} else {
-		opts.Select = func(item any) (any, error) {
-			setEnv(item)
-			return i.evalExpr(q.Select)
-		}
-	}
-
-	if q.Sort != nil {
-		opts.SortKey = func(item any) (any, error) {
-			setEnv(item)
-			return i.evalExpr(q.Sort)
-		}
-	}
-
-	if q.Skip != nil {
-		v, err := i.evalExpr(q.Skip)
-		if err != nil {
-			return nil, err
-		}
-		n, ok := v.(int)
-		if !ok {
-			return nil, fmt.Errorf("skip expects int, got %T", v)
-		}
-		opts.Skip = &n
-	}
-
-	if q.Take != nil {
-		v, err := i.evalExpr(q.Take)
-		if err != nil {
-			return nil, err
-		}
-		n, ok := v.(int)
-		if !ok {
-			return nil, fmt.Errorf("take expects int, got %T", v)
-		}
-		opts.Take = &n
-	}
-
-	return data.Query(list, opts)
+	return data.EvalQuery(src, q, child, func(e *parser.Expr) (any, error) {
+		return i.evalExpr(e)
+	})
 }
 
 func (i *Interpreter) evalMatch(m *parser.MatchExpr) (any, error) {
